Implement FETCH command handling and response building for IMAP server
- Added `handleFetchForUIDs` and `handleFetch` methods to manage FETCH requests for messages by UID and sequence number.
- Introduced middleware functions for authentication and mailbox selection checks.
- Created response builders for BODYSTRUCTURE and ENVELOPE according to RFC 3501 specifications.
- Enhanced message parsing to support multipart messages and various body formats.
- Implemented error handling for database interactions and message reconstruction.
- Refactored code for better readability and maintainability.
