+++ internal/server/handler_message.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"mime"
-	"mime/multipart"
 	"net"
 	"strconv"
 	"strings"
@@ -17,426 +15,6 @@ import (
 	"raven/internal/models"
 	"raven/internal/server/utils"
 )
-
-// ===== FETCH =====
-
-// handleFetchForUIDs handles FETCH for a list of UIDs (used by UID FETCH command)
-func (s *IMAPServer) handleFetchForUIDs(conn net.Conn, tag string, uids []int, items string, state *models.ClientState) {
-	// Get appropriate database (user or role mailbox)
-	targetDB, _, err := s.GetSelectedDB(state)
-	if err != nil {
-		return
-	}
-
-	for _, uid := range uids {
-		// Get message details by UID
-		var messageID int64
-		var seqNum int
-		var flags sql.NullString
-
-		err := targetDB.QueryRow(`
-			SELECT mm.message_id, mm.flags,
-				(SELECT COUNT(*) FROM message_mailbox mm2
-				 WHERE mm2.mailbox_id = mm.mailbox_id AND mm2.uid <= mm.uid) as seq_num
-			FROM message_mailbox mm
-			WHERE mm.mailbox_id = ? AND mm.uid = ?
-		`, state.SelectedMailboxID, uid).Scan(&messageID, &flags, &seqNum)
-
-		if err != nil {
-			// Non-existent UID is silently ignored
-			continue
-		}
-
-		// Process this message using the same logic as handleFetch
-		s.processFetchForMessage(conn, messageID, int64(uid), seqNum, flags.String, items, state)
-	}
-}
-
-func (s *IMAPServer) handleFetch(conn net.Conn, tag string, parts []string, state *models.ClientState) {
-	if !state.Authenticated {
-		s.sendResponse(conn, fmt.Sprintf("%s NO Please authenticate first", tag))
-		return
-	}
-
-	if state.SelectedMailboxID == 0 {
-		s.sendResponse(conn, fmt.Sprintf("%s NO No folder selected", tag))
-		return
-	}
-
-	if len(parts) < 4 {
-		s.sendResponse(conn, fmt.Sprintf("%s BAD FETCH requires sequence and items", tag))
-		return
-	}
-
-	// Get appropriate database (user or role mailbox)
-	targetDB, _, err := s.GetSelectedDB(state)
-	if err != nil {
-		s.sendResponse(conn, fmt.Sprintf("%s NO Database error", tag))
-		return
-	}
-
-	sequence := parts[2]
-	items := strings.Join(parts[3:], " ")
-
-	// Handle FETCH macros: ALL, FAST, FULL
-	itemsUpper := strings.ToUpper(strings.TrimSpace(items))
-	switch itemsUpper {
-	case "ALL":
-		items = "FLAGS INTERNALDATE RFC822.SIZE ENVELOPE"
-	case "FAST":
-		items = "FLAGS INTERNALDATE RFC822.SIZE"
-	case "FULL":
-		items = "FLAGS INTERNALDATE RFC822.SIZE ENVELOPE BODY"
-	default:
-		// Remove parentheses if present
-		items = strings.Trim(items, "()")
-	}
-
-	var rows *sql.Rows
-
-	// Support for sequence ranges (e.g., 1:2, 2:4, 1:*, *)
-	seqRange := strings.Split(sequence, ":")
-	var start, end int
-	var useRange bool
-
-	if len(seqRange) == 2 {
-		useRange = true
-		if seqRange[0] == "*" {
-			start = -1 // will handle below
-		} else {
-			start, err = strconv.Atoi(seqRange[0])
-			if err != nil || start < 1 {
-				s.sendResponse(conn, fmt.Sprintf("%s BAD Invalid sequence number", tag))
-				return
-			}
-		}
-		if seqRange[1] == "*" {
-			// Get max count for end using new schema
-			end, _ = db.GetMessageCountPerUser(targetDB, state.SelectedMailboxID)
-		} else {
-			end, err = strconv.Atoi(seqRange[1])
-			if err != nil || end < 1 {
-				s.sendResponse(conn, fmt.Sprintf("%s BAD Invalid sequence number", tag))
-				return
-			}
-		}
-		if start == -1 {
-			start = end
-		}
-		if end < start {
-			end = start
-		}
-		// Query message_mailbox for messages in selected mailbox using new schema
-		query := `SELECT mm.message_id, mm.uid, mm.flags
-		          FROM message_mailbox mm
-		          WHERE mm.mailbox_id = ?
-		          ORDER BY mm.uid ASC LIMIT ? OFFSET ?`
-		rows, err = targetDB.Query(query, state.SelectedMailboxID, end-start+1, start-1)
-	} else if sequence == "1:*" || sequence == "*" {
-		query := `SELECT mm.message_id, mm.uid, mm.flags
-		          FROM message_mailbox mm
-		          WHERE mm.mailbox_id = ?
-		          ORDER BY mm.uid ASC`
-		rows, err = targetDB.Query(query, state.SelectedMailboxID)
-	} else {
-		msgNum, parseErr := strconv.Atoi(sequence)
-		if parseErr != nil {
-			s.sendResponse(conn, fmt.Sprintf("%s BAD Invalid sequence number", tag))
-			return
-		}
-		query := `SELECT mm.message_id, mm.uid, mm.flags
-		          FROM message_mailbox mm
-		          WHERE mm.mailbox_id = ?
-		          ORDER BY mm.uid ASC LIMIT 1 OFFSET ?`
-		rows, err = targetDB.Query(query, state.SelectedMailboxID, msgNum-1)
-	}
-
-	if err != nil {
-		s.sendResponse(conn, fmt.Sprintf("%s NO Database error", tag))
-		return
-	}
-	defer rows.Close()
-
-	seqNum := 1
-	if useRange {
-		seqNum = start
-	}
-	for rows.Next() {
-		var messageID int64
-		var uid int64
-		var flagsStr sql.NullString
-		if err := rows.Scan(&messageID, &uid, &flagsStr); err != nil {
-			continue
-		}
-
-		flags := ""
-		if flagsStr.Valid {
-			flags = flagsStr.String
-		}
-
-		// Process this message
-		s.processFetchForMessage(conn, messageID, uid, seqNum, flags, items, state)
-		seqNum++
-	}
-
-	s.sendResponse(conn, fmt.Sprintf("%s OK FETCH completed", tag))
-}
-
-// processFetchForMessage processes a single message for FETCH/UID FETCH
-func (s *IMAPServer) processFetchForMessage(conn net.Conn, messageID, uid int64, seqNum int, flags, items string, state *models.ClientState) {
-	// Get appropriate database (user or role mailbox)
-	targetDB, _, err := s.GetSelectedDB(state)
-	if err != nil {
-		return
-	}
-
-	// Reconstruct message from new schema
-	rawMsg, err := parser.ReconstructMessage(targetDB, messageID)
-	if err != nil {
-		// If reconstruction fails, skip this message
-		return
-	}
-
-		if !strings.Contains(rawMsg, "\r\n") {
-			rawMsg = strings.ReplaceAll(rawMsg, "\n", "\r\n")
-		}
-
-		itemsUpper := strings.ToUpper(items)
-		responseParts := []string{}
-		var literalData string // Store literal data separately
-
-		if strings.Contains(itemsUpper, "UID") {
-			responseParts = append(responseParts, fmt.Sprintf("UID %d", uid))
-		}
-		if strings.Contains(itemsUpper, "FLAGS") {
-			if flags == "" {
-				flags = "()"
-			} else {
-				flags = fmt.Sprintf("(%s)", flags)
-			}
-			responseParts = append(responseParts, fmt.Sprintf("FLAGS %s", flags))
-		}
-		if strings.Contains(itemsUpper, "INTERNALDATE") {
-			var internalDate time.Time
-			// Query message_mailbox for internal_date using new schema
-			query := "SELECT internal_date FROM message_mailbox WHERE message_id = ? AND mailbox_id = ?"
-			err := targetDB.QueryRow(query, messageID, state.SelectedMailboxID).Scan(&internalDate)
-
-			var dateStr string
-			if err != nil || internalDate.IsZero() {
-				dateStr = "01-Jan-1970 00:00:00 +0000"
-			} else {
-				// Format as RFC 3501: "02-Jan-2006 15:04:05 -0700"
-				dateStr = internalDate.Format("02-Jan-2006 15:04:05 -0700")
-			}
-			responseParts = append(responseParts, fmt.Sprintf("INTERNALDATE \"%s\"", dateStr))
-		}
-		if strings.Contains(itemsUpper, "RFC822.SIZE") {
-			responseParts = append(responseParts, fmt.Sprintf("RFC822.SIZE %d", len(rawMsg)))
-		}
-
-		// Handle ENVELOPE
-		if strings.Contains(itemsUpper, "ENVELOPE") {
-			envelope := s.buildEnvelope(rawMsg)
-			responseParts = append(responseParts, envelope)
-		}
-
-		// Handle BODYSTRUCTURE
-		if strings.Contains(itemsUpper, "BODYSTRUCTURE") {
-			bodyStructure := s.buildBodyStructure(rawMsg)
-			responseParts = append(responseParts, bodyStructure)
-		}
-
-		// Handle BODY (non-extensible BODYSTRUCTURE)
-		if strings.Contains(itemsUpper, "BODY") && !strings.Contains(itemsUpper, "BODY[") && !strings.Contains(itemsUpper, "BODY.PEEK") && !strings.Contains(itemsUpper, "BODYSTRUCTURE") {
-			// BODY is the non-extensible form of BODYSTRUCTURE
-			bodyStructure := s.buildBodyStructure(rawMsg)
-			// Replace BODYSTRUCTURE with BODY in the response
-			bodyStructure = strings.Replace(bodyStructure, "BODYSTRUCTURE", "BODY", 1)
-			responseParts = append(responseParts, bodyStructure)
-		}
-
-		// Handle multiple body parts - process each separately
-		// Handle BODY.PEEK[HEADER.FIELDS (...)] or BODY[HEADER.FIELDS (...)] - specific header fields
-		if strings.Contains(itemsUpper, "BODY.PEEK[HEADER.FIELDS") || strings.Contains(itemsUpper, "BODY[HEADER.FIELDS") {
-			start := strings.Index(itemsUpper, "BODY.PEEK[HEADER.FIELDS")
-			if start == -1 {
-				start = strings.Index(itemsUpper, "BODY[HEADER.FIELDS")
-			}
-
-			// Extract requested header field names
-			requestedHeaders := []string{"FROM", "TO", "CC", "BCC", "SUBJECT", "DATE", "MESSAGE-ID", "PRIORITY", "X-PRIORITY", "REFERENCES", "NEWSGROUPS", "IN-REPLY-TO", "CONTENT-TYPE", "REPLY-TO"}
-			if start != -1 {
-				isPeek := strings.Contains(itemsUpper, "BODY.PEEK[HEADER.FIELDS")
-				prefixLen := len("BODY[HEADER.FIELDS (")
-				if isPeek {
-					prefixLen = len("BODY.PEEK[HEADER.FIELDS (")
-				}
-
-				fieldsStr := items[start+prefixLen:]
-				closeParen := strings.Index(fieldsStr, ")")
-				if closeParen != -1 {
-					fieldsStr = fieldsStr[:closeParen]
-					fields := strings.Fields(fieldsStr)
-					if len(fields) > 0 {
-						requestedHeaders = []string{}
-						for _, f := range fields {
-							requestedHeaders = append(requestedHeaders, strings.ToUpper(strings.TrimSpace(f)))
-						}
-					}
-				}
-			}
-
-			// Extract only the requested headers from the message
-			headersMap := map[string]string{}
-			lines := strings.Split(rawMsg, "\r\n")
-			currentHeader := ""
-			for _, line := range lines {
-				if line == "" {
-					break // End of headers
-				}
-				// Check if this is a continuation line (starts with space or tab)
-				if len(line) > 0 && (line[0] == ' ' || line[0] == '\t') {
-					if currentHeader != "" {
-						headersMap[currentHeader] += "\r\n" + line
-					}
-					continue
-				}
-				// New header line
-				colonIdx := strings.Index(line, ":")
-				if colonIdx != -1 {
-					headerName := strings.ToUpper(strings.TrimSpace(line[:colonIdx]))
-					for _, h := range requestedHeaders {
-						if headerName == h {
-							currentHeader = h
-							headersMap[h] = line
-							break
-						}
-					}
-				}
-			}
-
-			// Build response with requested headers in order
-			var headerLines []string
-			for _, h := range requestedHeaders {
-				if val, ok := headersMap[h]; ok {
-					headerLines = append(headerLines, val)
-				}
-			}
-			headersStr := strings.Join(headerLines, "\r\n")
-			if len(headersStr) > 0 {
-				headersStr += "\r\n"
-			}
-			headersStr += "\r\n" // Final blank line
-				// Match the exact format the client requested
-		fieldList := strings.Join(requestedHeaders, " ")
-		responseParts = append(responseParts, fmt.Sprintf("BODY[HEADER.FIELDS (%s)]", fieldList))
-		literalData = fmt.Sprintf("{%d}\r\n%s", len(headersStr), headersStr)
-	}
-
-	// Handle BODY.PEEK[TEXT] or BODY[TEXT] - message body only (can be combined with other parts)
-	if strings.Contains(itemsUpper, "BODY.PEEK[TEXT]") || strings.Contains(itemsUpper, "BODY[TEXT]") {
-		headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-		body := ""
-		if headerEnd != -1 {
-			body = rawMsg[headerEnd+4:] // skip the double CRLF
-		}
-
-		// Check for partial fetch like BODY.PEEK[TEXT]<0.2048>
-		partialStart := 0
-		partialLength := len(body)
-		if strings.Contains(itemsUpper, "<") && strings.Contains(itemsUpper, ">") {
-			startIdx := strings.Index(itemsUpper, "<")
-			endIdx := strings.Index(itemsUpper, ">")
-			if startIdx != -1 && endIdx > startIdx {
-				partialSpec := itemsUpper[startIdx+1:endIdx]
-				fmt.Sscanf(partialSpec, "%d.%d", &partialStart, &partialLength)
-				if partialStart < len(body) {
-					endPos := partialStart + partialLength
-					if endPos > len(body) {
-						endPos = len(body)
-					}
-					body = body[partialStart:endPos]
-				} else {
-					body = ""
-				}
-			}
-		}
-
-		if literalData != "" {
-			literalData += " "
-		}
-		responseParts = append(responseParts, "BODY[TEXT]")
-		literalData += fmt.Sprintf("{%d}\r\n%s", len(body), body)
-	}
-
-	// Handle BODY.PEEK[HEADER] or BODY[HEADER] - all headers (check it's not HEADER.FIELDS)
-	if (strings.Contains(itemsUpper, "BODY.PEEK[HEADER]") || strings.Contains(itemsUpper, "BODY[HEADER]")) &&
-	   !strings.Contains(itemsUpper, "HEADER.FIELDS") {
-		headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-		headers := rawMsg
-		if headerEnd != -1 {
-			headers = rawMsg[:headerEnd+2] // include last CRLF
-		}
-		if literalData != "" {
-			literalData += " "
-		}
-		responseParts = append(responseParts, "BODY[HEADER]")
-		literalData += fmt.Sprintf("{%d}\r\n%s", len(headers), headers)
-	}
-
-	// Handle RFC822.HEADER - return only the header portion
-	if strings.Contains(itemsUpper, "RFC822.HEADER") {
-		headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-		headers := rawMsg
-		if headerEnd != -1 {
-			headers = rawMsg[:headerEnd+2] // include last CRLF
-		}
-		if literalData != "" {
-			literalData += " "
-		}
-		responseParts = append(responseParts, "RFC822.HEADER")
-		literalData += fmt.Sprintf("{%d}\r\n%s", len(headers), headers)
-	}
-
-	// Handle RFC822.TEXT - body text only (excluding headers)
-	if strings.Contains(itemsUpper, "RFC822.TEXT") {
-		headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-		body := ""
-		if headerEnd != -1 {
-			body = rawMsg[headerEnd+4:] // skip the double CRLF
-		}
-		if literalData != "" {
-			literalData += " "
-		}
-		responseParts = append(responseParts, "RFC822.TEXT")
-		literalData += fmt.Sprintf("{%d}\r\n%s", len(body), body)
-	}
-
-	// Handle BODY[] / BODY.PEEK[] / RFC822 / RFC822.PEEK - full message
-	if strings.Contains(itemsUpper, "BODY[]") || strings.Contains(itemsUpper, "BODY.PEEK[]") ||
-	   strings.Contains(itemsUpper, "RFC822.PEEK") ||
-	   (strings.Contains(itemsUpper, "RFC822") && !strings.Contains(itemsUpper, "RFC822.SIZE") &&
-	    !strings.Contains(itemsUpper, "RFC822.HEADER") && !strings.Contains(itemsUpper, "RFC822.TEXT") && !strings.Contains(itemsUpper, "RFC822.PEEK")) {
-		if literalData != "" {
-			literalData += " "
-		}
-		responseParts = append(responseParts, "BODY[]")
-		literalData += fmt.Sprintf("{%d}\r\n%s", len(rawMsg), rawMsg)
-	}
-
-	if len(responseParts) > 0 {
-		responseStr := fmt.Sprintf("* %d FETCH (%s", seqNum, strings.Join(responseParts, " "))
-		if literalData != "" {
-			responseStr += " " + literalData + ")"
-		} else {
-			responseStr += ")"
-		}
-		s.sendResponse(conn, responseStr)
-	} else {
-		s.sendResponse(conn, fmt.Sprintf("* %d FETCH (FLAGS ())", seqNum))
-	}
-}
 
 // ===== SEARCH =====
 
@@ -1780,491 +1358,4 @@ func (s *IMAPServer) handleCheck(conn net.Conn, tag string, state *models.Client
 // HandleCheck exports the check handler for testing
 func (s *IMAPServer) HandleCheck(conn net.Conn, tag string, state *models.ClientState) {
 	s.handleCheck(conn, tag, state)
-}
-
-// ===== HELPER METHODS FOR FETCH =====
-
-// buildEnvelope builds an ENVELOPE structure from a message
-// ENVELOPE: (date subject from sender reply-to to cc bcc in-reply-to message-id)
-func (s *IMAPServer) buildEnvelope(rawMsg string) string {
-	// Extract headers
-	date := s.extractHeader(rawMsg, "Date")
-	subject := s.extractHeader(rawMsg, "Subject")
-	from := s.extractHeader(rawMsg, "From")
-	sender := s.extractHeader(rawMsg, "Sender")
-	replyTo := s.extractHeader(rawMsg, "Reply-To")
-	to := s.extractHeader(rawMsg, "To")
-	cc := s.extractHeader(rawMsg, "Cc")
-	bcc := s.extractHeader(rawMsg, "Bcc")
-	inReplyTo := s.extractHeader(rawMsg, "In-Reply-To")
-	messageID := s.extractHeader(rawMsg, "Message-ID")
-
-	// If sender is empty, use from
-	if sender == "" {
-		sender = from
-	}
-	// If reply-to is empty, use from
-	if replyTo == "" {
-		replyTo = from
-	}
-
-	// Build ENVELOPE structure
-	envelope := fmt.Sprintf("ENVELOPE (%s %s %s %s %s %s %s %s %s %s)",
-		s.quoteOrNIL(date),
-		s.quoteOrNIL(subject),
-		s.parseAddressList(from),
-		s.parseAddressList(sender),
-		s.parseAddressList(replyTo),
-		s.parseAddressList(to),
-		s.parseAddressList(cc),
-		s.parseAddressList(bcc),
-		s.quoteOrNIL(inReplyTo),
-		s.quoteOrNIL(messageID),
-	)
-
-	return envelope
-}
-
-// extractHeader extracts a header value from a raw message
-func (s *IMAPServer) extractHeader(rawMsg string, headerName string) string {
-	lines := strings.Split(rawMsg, "\n")
-	headerNameUpper := strings.ToUpper(headerName)
-	var headerValue strings.Builder
-	inHeader := false
-
-	for _, line := range lines {
-		line = strings.TrimRight(line, "\r")
-		if line == "" {
-			break // End of headers
-		}
-
-		// Check if continuation line
-		if len(line) > 0 && (line[0] == ' ' || line[0] == '\t') {
-			if inHeader {
-				headerValue.WriteString(" ")
-				headerValue.WriteString(strings.TrimSpace(line))
-			}
-			continue
-		}
-
-		// New header line
-		colonIdx := strings.Index(line, ":")
-		if colonIdx != -1 {
-			currentHeader := strings.TrimSpace(line[:colonIdx])
-			if strings.ToUpper(currentHeader) == headerNameUpper {
-				inHeader = true
-				headerValue.WriteString(strings.TrimSpace(line[colonIdx+1:]))
-			} else {
-				inHeader = false
-			}
-		}
-	}
-
-	return headerValue.String()
-}
-
-// quoteOrNIL quotes a string or returns NIL if empty
-func (s *IMAPServer) quoteOrNIL(str string) string {
-	if str == "" {
-		return "NIL"
-	}
-	// Escape quotes and backslashes in the string
-	str = strings.ReplaceAll(str, "\\", "\\\\")
-	str = strings.ReplaceAll(str, "\"", "\\\"")
-	return fmt.Sprintf("\"%s\"", str)
-}
-
-// parseAddressList parses an address header into IMAP address list format
-// Address list: ((name route mailbox host) ...) or NIL
-func (s *IMAPServer) parseAddressList(addresses string) string {
-	if addresses == "" {
-		return "NIL"
-	}
-
-	// Simple parser - split by comma
-	addrs := strings.Split(addresses, ",")
-	var addrStructs []string
-
-	for _, addr := range addrs {
-		addr = strings.TrimSpace(addr)
-		if addr == "" {
-			continue
-		}
-
-		// Parse "Name <email@domain.com>" or just "email@domain.com"
-		name := ""
-		email := addr
-
-		// Check for name part
-		if strings.Contains(addr, "<") && strings.Contains(addr, ">") {
-			start := strings.Index(addr, "<")
-			end := strings.Index(addr, ">")
-			name = strings.TrimSpace(addr[:start])
-			email = addr[start+1 : end]
-			// Remove quotes from name if present
-			name = strings.Trim(name, "\"")
-		}
-
-		// Parse email into mailbox@host
-		mailbox := email
-		host := ""
-		if strings.Contains(email, "@") {
-			parts := strings.SplitN(email, "@", 2)
-			mailbox = parts[0]
-			host = parts[1]
-		}
-
-		// Build address structure: (name route mailbox host)
-		// route is always NIL in modern email
-		addrStruct := fmt.Sprintf("(%s NIL %s %s)",
-			s.quoteOrNIL(name),
-			s.quoteOrNIL(mailbox),
-			s.quoteOrNIL(host),
-		)
-		addrStructs = append(addrStructs, addrStruct)
-	}
-
-	if len(addrStructs) == 0 {
-		return "NIL"
-	}
-
-	return "(" + strings.Join(addrStructs, " ") + ")"
-}
-
-// buildBodyStructure builds a BODYSTRUCTURE response
-func (s *IMAPServer) buildBodyStructure(rawMsg string) string {
-	// Extract Content-Type header
-	contentType := s.extractHeader(rawMsg, "Content-Type")
-	if contentType == "" {
-		contentType = "text/plain; charset=us-ascii"
-	}
-
-	// Parse content type and parameters
-	mediaType, params, err := mime.ParseMediaType(contentType)
-	if err != nil {
-		mediaType = "text/plain"
-		params = map[string]string{"charset": "us-ascii"}
-	}
-
-	// Split media type into type and subtype
-	typeParts := strings.SplitN(mediaType, "/", 2)
-	mainType := "TEXT"
-	subType := "PLAIN"
-	if len(typeParts) == 2 {
-		mainType = strings.ToUpper(typeParts[0])
-		subType = strings.ToUpper(typeParts[1])
-	}
-
-	// Handle multipart messages
-	if strings.HasPrefix(strings.ToLower(mediaType), "multipart/") {
-		boundary := params["boundary"]
-		if boundary != "" {
-			return s.buildMultipartBodyStructure(rawMsg, mainType, subType, boundary)
-		}
-	}
-
-	// For non-multipart messages, return basic body structure
-	// Get message body
-	headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-	if headerEnd == -1 {
-		headerEnd = strings.Index(rawMsg, "\n\n")
-	}
-	body := ""
-	if headerEnd != -1 {
-		body = rawMsg[headerEnd+4:]
-	}
-
-	// Get encoding
-	encoding := s.extractHeader(rawMsg, "Content-Transfer-Encoding")
-	if encoding == "" {
-		encoding = "7BIT"
-	}
-	encoding = strings.ToUpper(encoding)
-
-	// Build parameters list
-	paramList := s.buildParamList(params)
-
-	// Get Content-ID and Content-Description
-	contentID := s.extractHeader(rawMsg, "Content-ID")
-	contentDesc := s.extractHeader(rawMsg, "Content-Description")
-
-	// Count lines for text types
-	lines := 0
-	if mainType == "TEXT" {
-		lines = strings.Count(body, "\n")
-	}
-
-	// Format: (type subtype (params) id description encoding size [lines for text])
-	if mainType == "TEXT" {
-		return fmt.Sprintf("BODYSTRUCTURE (%s %s %s %s %s %s %d %d)",
-			s.quoteOrNIL(mainType),
-			s.quoteOrNIL(subType),
-			paramList,
-			s.quoteOrNIL(contentID),
-			s.quoteOrNIL(contentDesc),
-			s.quoteOrNIL(encoding),
-			len(body),
-			lines,
-		)
-	}
-
-	return fmt.Sprintf("BODYSTRUCTURE (%s %s %s %s %s %s %d)",
-		s.quoteOrNIL(mainType),
-		s.quoteOrNIL(subType),
-		paramList,
-		s.quoteOrNIL(contentID),
-		s.quoteOrNIL(contentDesc),
-		s.quoteOrNIL(encoding),
-		len(body),
-	)
-}
-
-// buildMultipartBodyStructure builds BODYSTRUCTURE for multipart messages
-func (s *IMAPServer) buildMultipartBodyStructure(rawMsg, mainType, subType, boundary string) string {
-	// Get the body part (after headers)
-	headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-	if headerEnd == -1 {
-		headerEnd = strings.Index(rawMsg, "\n\n")
-		if headerEnd == -1 {
-			// No headers found, return fallback
-			return s.buildFallbackBodyStructure(mainType, subType)
-		}
-		headerEnd += 2
-	} else {
-		headerEnd += 4
-	}
-	body := rawMsg[headerEnd:]
-
-	// Normalize line endings for multipart parsing
-	if !strings.Contains(body, "\r\n") {
-		body = strings.ReplaceAll(body, "\n", "\r\n")
-	}
-
-	// Parse multipart body
-	var parts []string
-	mr := multipart.NewReader(strings.NewReader(body), boundary)
-
-	for {
-		part, err := mr.NextPart()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			// Log error but try to continue parsing (might be a boundary issue)
-			// Fall back to manual parsing if multipart.Reader fails completely
-			if len(parts) == 0 {
-				return s.buildFallbackMultipartBodyStructure(rawMsg, mainType, subType, boundary)
-			}
-			break
-		}
-
-		// Read part content
-		partContent, err := io.ReadAll(part)
-		if err != nil {
-			continue
-		}
-
-		// Build part headers
-		var partHeaders strings.Builder
-		for key, values := range part.Header {
-			for _, value := range values {
-				partHeaders.WriteString(fmt.Sprintf("%s: %s\r\n", key, value))
-			}
-		}
-		partHeaders.WriteString("\r\n")
-		fullPart := partHeaders.String() + string(partContent)
-
-		// Build BODYSTRUCTURE for this part
-		partStruct := s.buildPartStructure(fullPart)
-		parts = append(parts, partStruct)
-	}
-
-	if len(parts) == 0 {
-		// Fallback to manual parsing if multipart.Reader failed
-		return s.buildFallbackMultipartBodyStructure(rawMsg, mainType, subType, boundary)
-	}
-
-	// Multipart BODYSTRUCTURE format: BODYSTRUCTURE ((part1)(part2)... subtype)
-	// Note: Each part is already a complete structure without BODYSTRUCTURE keyword
-	return fmt.Sprintf("BODYSTRUCTURE (%s %s)", strings.Join(parts, " "), s.quoteOrNIL(subType))
-}
-
-// buildPartStructure builds BODYSTRUCTURE for a single MIME part
-func (s *IMAPServer) buildPartStructure(partMsg string) string {
-	// Extract Content-Type
-	contentType := s.extractHeader(partMsg, "Content-Type")
-	if contentType == "" {
-		contentType = "text/plain; charset=us-ascii"
-	}
-
-	// Parse media type
-	mediaType, params, err := mime.ParseMediaType(contentType)
-	if err != nil {
-		mediaType = "text/plain"
-		params = map[string]string{"charset": "us-ascii"}
-	}
-
-	typeParts := strings.SplitN(mediaType, "/", 2)
-	mainType := "TEXT"
-	subType := "PLAIN"
-	if len(typeParts) == 2 {
-		mainType = strings.ToUpper(typeParts[0])
-		subType = strings.ToUpper(typeParts[1])
-	}
-
-	// Get encoding
-	encoding := s.extractHeader(partMsg, "Content-Transfer-Encoding")
-	if encoding == "" {
-		encoding = "7BIT"
-	}
-	encoding = strings.ToUpper(encoding)
-
-	// Get body
-	headerEnd := strings.Index(partMsg, "\r\n\r\n")
-	if headerEnd == -1 {
-		headerEnd = strings.Index(partMsg, "\n\n")
-		if headerEnd == -1 {
-			headerEnd = 0
-		} else {
-			headerEnd += 2
-		}
-	} else {
-		headerEnd += 4
-	}
-	body := ""
-	if headerEnd < len(partMsg) {
-		body = partMsg[headerEnd:]
-	}
-
-	// Build parameters
-	paramList := s.buildParamList(params)
-
-	// Get Content-ID and Content-Description
-	contentID := s.extractHeader(partMsg, "Content-ID")
-	contentDesc := s.extractHeader(partMsg, "Content-Description")
-
-	// Get Content-Disposition and filename
-	disposition := s.extractHeader(partMsg, "Content-Disposition")
-	var dispList string
-	if disposition != "" {
-		dispType, dispParams, _ := mime.ParseMediaType(disposition)
-		dispParamList := s.buildParamList(dispParams)
-		dispList = fmt.Sprintf("(%s %s)", s.quoteOrNIL(strings.ToUpper(dispType)), dispParamList)
-	} else {
-		dispList = "NIL"
-	}
-
-	// Count lines for text types
-	lines := 0
-	if mainType == "TEXT" {
-		lines = strings.Count(body, "\n")
-		return fmt.Sprintf("(%s %s %s %s %s %s %d %d NIL %s NIL)",
-			s.quoteOrNIL(mainType),
-			s.quoteOrNIL(subType),
-			paramList,
-			s.quoteOrNIL(contentID),
-			s.quoteOrNIL(contentDesc),
-			s.quoteOrNIL(encoding),
-			len(body),
-			lines,
-			dispList,
-		)
-	}
-
-	return fmt.Sprintf("(%s %s %s %s %s %s %d NIL %s NIL)",
-		s.quoteOrNIL(mainType),
-		s.quoteOrNIL(subType),
-		paramList,
-		s.quoteOrNIL(contentID),
-		s.quoteOrNIL(contentDesc),
-		s.quoteOrNIL(encoding),
-		len(body),
-		dispList,
-	)
-}
-
-// buildParamList builds parameter list for BODYSTRUCTURE
-func (s *IMAPServer) buildParamList(params map[string]string) string {
-	if len(params) == 0 {
-		return "NIL"
-	}
-
-	var paramPairs []string
-	for key, value := range params {
-		paramPairs = append(paramPairs, fmt.Sprintf("%s %s",
-			s.quoteOrNIL(strings.ToUpper(key)),
-			s.quoteOrNIL(value)))
-	}
-
-	return fmt.Sprintf("(%s)", strings.Join(paramPairs, " "))
-}
-
-// buildFallbackBodyStructure builds a simple fallback BODYSTRUCTURE
-func (s *IMAPServer) buildFallbackBodyStructure(mainType, subType string) string {
-	return fmt.Sprintf("BODYSTRUCTURE (%s %s NIL NIL NIL \"7BIT\" 0)",
-		s.quoteOrNIL(mainType), s.quoteOrNIL(subType))
-}
-
-// buildFallbackMultipartBodyStructure manually parses multipart messages when multipart.Reader fails
-func (s *IMAPServer) buildFallbackMultipartBodyStructure(rawMsg, mainType, subType, boundary string) string {
-	// Get the body part (after headers)
-	headerEnd := strings.Index(rawMsg, "\r\n\r\n")
-	if headerEnd == -1 {
-		headerEnd = strings.Index(rawMsg, "\n\n")
-		if headerEnd == -1 {
-			return s.buildFallbackBodyStructure(mainType, subType)
-		}
-		headerEnd += 2
-	} else {
-		headerEnd += 4
-	}
-	body := rawMsg[headerEnd:]
-
-	// Normalize line endings
-	if !strings.Contains(body, "\r\n") {
-		body = strings.ReplaceAll(body, "\n", "\r\n")
-	}
-
-	// Manually split by boundary
-	boundaryMarker := "--" + boundary
-	closeBoundary := "--" + boundary + "--"
-
-	// Split the body into parts
-	partSections := strings.Split(body, boundaryMarker)
-	var parts []string
-
-	for i, section := range partSections {
-		// Skip the preamble (before first boundary) and epilogue (after final boundary)
-		if i == 0 || strings.TrimSpace(section) == "" {
-			continue
-		}
-
-		// Check if this is the closing boundary
-		if strings.HasPrefix(strings.TrimSpace(section), "--") {
-			break
-		}
-
-		// Remove the closing boundary if present
-		section = strings.TrimPrefix(section, "\r\n")
-		section = strings.TrimPrefix(section, "\n")
-
-		// Check if this section ends with the closing boundary
-		if idx := strings.Index(section, closeBoundary); idx != -1 {
-			section = section[:idx]
-		}
-
-		// Parse this part's structure
-		if strings.TrimSpace(section) != "" {
-			partStruct := s.buildPartStructure(section)
-			parts = append(parts, partStruct)
-		}
-	}
-
-	if len(parts) == 0 {
-		// Still no parts found, return simple fallback
-		return s.buildFallbackBodyStructure(mainType, subType)
-	}
-
-	// Return the multipart structure
-	return fmt.Sprintf("BODYSTRUCTURE (%s %s)", strings.Join(parts, " "), s.quoteOrNIL(subType))
 }
